Yet another warning down (pps... I choose to cast rc since i is being   doubled... I expect the mismatched types were deliberate.)
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@86529 13f79535-47bb-0310-9956-ffa450edef68

+++ srclib/pcre/pcreposix.c
@@ -247,7 +247,7 @@ if (rc == 0) rc = nmatch;    /* All captured slots were filled in */
 if (rc >= 0)
   {
   size_t i;
-  for (i = 0; i < rc; i++)
+  for (i = 0; i < (size_t) rc; i++)
     {
     pmatch[i].rm_so = ovector[i*2];
     pmatch[i].rm_eo = ovector[i*2+1];
